Extracted Chat component from index.tsx
